Cleanups for I3DM to GLB
Split PNTS/B3DM/I3DM migration into dedicated classes.
Tried to get the rotation right in any many cases as
possible...
